pyproject.toml
[project]
name = "spiritual-direction-chat"
version = "0.1.0"
description = "RAG-powered spiritual director chat built with FastAPI, Chroma, and LM Studio."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "chromadb>=1.3.5",
  "fastapi>=0.122.0",
  "httpx>=0.28.1",
  "pydantic>=2.12.4",
  "python-dotenv>=1.2.1",
  "uvicorn[standard]>=0.38.0",
]

[project.urls]
Homepage = "https://github.com/your-username/spiritual-direction-chat"
Repository = "https://github.com/your-username/spiritual-direction-chat"

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "ruff>=0.14.6",
]

[tool.ruff]
line-length = 120
target-version = "py310"
indent-width = 2

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
